edited timer

+++ shared/src/main/scala/poolmate/Entity.scala
@@ -73,7 +73,7 @@ final case class Pump(id: Long = 0,
 
 final case class Timer(id: Long = 0,
                        poolId: Long = 0,
-                       installed: Int = 0,
+                       installed: Long = 0,
                        model: String = "",
                        cost: Int = 0) extends Entity:
   def display = model
